requirements.txt
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
matplotlib==3.8.3
scikit-image==0.22.0
